refactor(coral): Move mockedEnvironmentResponse to environment-test-helper.ts
